FIX: Revert "FIX: Automatic auth flow with full page login/signup (#30928)"
This reverts commit 13f86c99eaa1eae781ad8c869fcea515436eade8.
