+++ app/assets/javascripts/discourse/app/instance-initializers/auth-complete.js
@@ -140,18 +140,14 @@ export default {
               };
 
               if (siteSettings.full_page_login) {
-                router
-                  .transitionTo("signup", {
-                    queryParams: { authComplete: true },
-                  })
-                  .then((signup) => {
-                    const signupController =
-                      signup.controller || owner.lookup("controller:signup");
-                    Object.keys(createAccountProps || {}).forEach((key) => {
-                      signupController.set(key, createAccountProps[key]);
-                    });
-                    signupController.handleSkipConfirmation();
+                router.transitionTo("signup").then((signup) => {
+                  const signupController =
+                    signup.controller || owner.lookup("controller:signup");
+                  Object.keys(createAccountProps || {}).forEach((key) => {
+                    signupController.set(key, createAccountProps[key]);
                   });
+                  signupController.handleSkipConfirmation();
+                });
               } else {
                 modal.show(CreateAccount, { model: createAccountProps });
               }
